// frontend/src/client/types.gen.ts
export type NoteItem = {
    title: string;
    content: string;
    flag: number;
    image?: (string | null);
};

export type NotePublic = {
    id: string;
    title: string;
    note_data: Array<NoteItem>;
    user_id: string;
    created_at?: (string | null);
    updated_at?: (string | null);
};